gajal-platform/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
Flask_JWT_Extended==4.6.0
Flask_PyMongo==2.3.0
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==20.1.0
sendgrid
secure-smtplib
